requirements_2.txt
--find-links https://download.pytorch.org/whl/cpu
FastAPI==0.111.0
Flask==3.0.3
Pillow==10.4.0
SQLAlchemy==2.0.31
Werkzeug==3.0.3
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
autopep8
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu
fastapi-cli==0.0.4
fonttools==4.53.1
frozenlist==1.4.1
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hugchat==0.4.9
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-experimental==0.0.62
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langcodes==3.4.0
langsmith==0.1.85
language-data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
marshmallow==3.21.3
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.22.0
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==1.5.3
pip_search
pygobject
playwright==1.45.0
playsound
plotly==5.22.0
pluggy==1.5.0
preshed==3.0.9
protobuf==5.27.2
PyAudio
pyarrow==16.1.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pyee==11.1.0
pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.11.4
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition
srsly==2.4.8
starlette==0.37.2
streamlit==1.36.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tornado==6.4.1
tts
tqdm==4.66.4
transformers==4.42.3
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wandb
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4